Pull out non-util functionality from observable grouping module
