+++ test/utils/test_observable_grouping.py
@@ -17,7 +17,7 @@ import numpy as np
 from qiskit.quantum_info import Pauli, PauliList
 from qiskit.circuit import QuantumCircuit, ClassicalRegister
 
-from circuit_knitting_toolbox.circuit_cutting.qpd.utils.observable_estimation import *
+from circuit_knitting_toolbox.utils.observable_grouping import *
 
 
 @ddt
@@ -87,56 +87,6 @@ class TestEstimatorUtilities(unittest.TestCase):
         assert observables_restricted_to_subsystem(qubits, list(observables)) == list(
             expected
         )
-
-    def test_append_measurement_circuit(self):
-        qc = self.qc1.copy()
-        with self.subTest("In place"):
-            qcx = qc.copy()
-            assert append_measurement_circuit(qcx, self.cog, inplace=True) is qcx
-        with self.subTest("Out of place"):
-            assert append_measurement_circuit(qc, self.cog) is not qc
-        with self.subTest("Correct measurement circuit"):
-            qc2 = self.qc2.copy()
-            qc2.measure(0, 1)
-            qc2.h(1)
-            qc2.measure(1, 2)
-            assert append_measurement_circuit(qc, self.cog) == qc2
-        with self.subTest("Mismatch between qubit_locations and number of qubits"):
-            with pytest.raises(ValueError) as e_info:
-                append_measurement_circuit(qc, self.cog, qubit_locations=[0])
-            assert (
-                e_info.value.args[0]
-                == "qubit_locations has 1 element(s) but the observable(s) have 2 qubit(s)."
-            )
-        with self.subTest("Mismatched qubits, no qubit_locations provided"):
-            cog = CommutingObservableGroup(Pauli("X"), [Pauli("X")])
-            with pytest.raises(ValueError) as e_info:
-                append_measurement_circuit(qc, cog)
-            assert (
-                e_info.value.args[0]
-                == "Quantum circuit qubit count (2) does not match qubit count of observable(s) (1).  Try providing `qubit_locations` explicitly."
-            )
-
-    @data(
-        ("000", [1, 1, 1]),
-        ("001", [-1, -1, -1]),
-        ("010", [-1, 1, -1]),
-        ("011", [1, -1, 1]),
-        ("100", [1, -1, -1]),
-        ("101", [-1, 1, 1]),
-        ("110", [-1, -1, 1]),
-        ("111", [1, 1, -1]),
-    )
-    @unpack
-    def test_process_outcome(self, outcome, expected):
-        num_qpd_bits = len(self.qc2.cregs[-2])
-        for o in (
-            outcome,
-            f"0b{outcome}",
-            int(f"0b{outcome}", 0),
-            hex(int(f"0b{outcome}", 0)),
-        ):
-            assert np.all(process_outcome(num_qpd_bits, self.cog, o) == expected)
 
     def test_cog_with_nontrivial_phase(self):
         with pytest.raises(ValueError) as e_info:
